Ajout wishlist endpoint

+++ src/auth/auth.service.ts
@@ -2,9 +2,10 @@ import { Injectable } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
 import { JwtService } from '@nestjs/jwt';
 import { UsersService } from '../users/users.service';
-import { Prisma, User } from '@prisma/client';
 import { LoginResponseDto } from 'src/users/dto/login-response.dto';
-import { UserResponseDto } from 'src/users/dto/user-response.dto';
+import { UserEntity } from 'src/users/entities/user.entity';
+import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { UserToken } from './interfaces/user-token.interface';
 
 @Injectable()
 export class AuthService {
@@ -13,16 +14,14 @@ export class AuthService {
     private jwtService: JwtService,
   ) {}
 
-  async login(user: User): Promise<LoginResponseDto> {
+  async login(user: UserToken): Promise<LoginResponseDto> {
     const payload = { email: user.email, sub: user.id };
     return {
       access_token: this.jwtService.sign(payload),
     };
   }
 
-  async createAccount(
-    userCreateInput: Prisma.UserCreateInput,
-  ): Promise<UserResponseDto> {
+  async createAccount(userCreateInput: CreateUserDto): Promise<UserEntity> {
     const saltOrRounds = 10;
     const hashedPassword = await bcrypt.hash(
       userCreateInput.password,
@@ -35,19 +34,18 @@ export class AuthService {
     return { id, email, username };
   }
 
-  async deleteAccount(user: User): Promise<UserResponseDto> {
-    const { id, email, username } = await this.usersService.deleteOne({
-      id: user.id,
-    });
+  async deleteAccount(userId: number): Promise<UserEntity> {
+    const { id, email, username } = await this.usersService.deleteOne(userId);
     return { id, email, username };
   }
 
-  async validateUser(email: string, password: string): Promise<User> {
-    const user = await this.usersService.findOne({ email });
+  async validateUser(email: string, password: string): Promise<UserEntity> {
+    const user = await this.usersService.findOne(email);
     if (!user) return null;
     const passwordValid = await bcrypt.compare(password, user.password);
     if (user && passwordValid) {
-      return user;
+      const { id, username, email } = user;
+      return { id, username, email };
     }
     return null;
   }
